Add files via upload
This script creates a SongWriter with a list of MusicPiece objects. Each MusicPiece has a title, a creation date, and a list of alleged infringements. Each alleged infringement includes the name of the infringer, the date of the alleged infringement, and any additional details. You can create a new MusicPiece and report infringements to it using the create_piece and report_infringement methods, and you can list all works and their alleged infringements with the list_works method.